[fix] segv related to disconeted

+++ src/Utils/Protocol/Protocol.hpp
@@ -215,9 +215,8 @@ namespace Utils
                 std::size_t args_byte_position = 0x0;
 
                 std::cout << "ACK writing msg" << std::endl;
-                std::cout << "InfoType: " << static_cast<uint8_t>(info) << std::endl;
-                std::cout << "InfoFunction: " << static_cast<uint8_t>(functionDefiner) << std::endl;
-
+                std::cout << "InfoType: " << static_cast<uint32_t>(static_cast<uint8_t>(info)) << std::endl;
+                std::cout << "InfoFunction: " << static_cast<uint32_t>(static_cast<uint8_t>(functionDefiner)) << std::endl;
 
                 appendFixedSizeTypeIntoBytes(msg, ack, true);
                 appendFixedSizeTypeIntoBytes(msg, info, false);
